RED-52829 add support to read from file system.
